// Ref/Web/ReferenceDataUpdateService.Web/src/models/RefCusCodeListUserView.ts
//------------------------------------------------------------------------------
// <auto-generated>
//     This code was generated from a template.
//
//     Manual changes to this file may cause unexpected behavior in your application.
//     Manual changes to this file will be overwritten if the code is regenerated.
// </auto-generated>
//------------------------------------------------------------------------------
import { IEntity } from "./IEntity";

export class RefCusCodeListUserView implements IEntity {
	ZZD_PK: string;
	ZZD_CountryOrGrouping: string;
	static readonly ZZD_CountryOrGrouping_MaxLength: number = 3;
	ZZD_CodeType: string;
	static readonly ZZD_CodeType_MaxLength: number = 5;
	ZZD_Code: string;
	static readonly ZZD_Code_MaxLength: number = 35;
	ZZD_Description: string;
	static readonly ZZD_Description_MaxLength: number = 2000;
	ZZD_StartDate: string;
	ZZD_EndDate: string;
	ZZD_IsAir: boolean;
	ZZD_IsSea: boolean;
	ZZD_IsFix: boolean;
	ZZD_IsRai: boolean;
	ZZD_IsRoa: boolean;
	ZZD_IsMai: boolean;
	ZZD_IsInw: boolean;
	ZZD_IsSystem: boolean;
	ZZD_IsPublished: boolean;
	ZZD_IsEditable: boolean;
}